[EFL] [GTK] [QT] fast/viewport/viewport-91.html is failing after r121555 https://bugs.webkit.org/show_bug.cgi?id=90286
Patch by Konrad Piascik <kpiascik@rim.com> on 2012-07-01
Reviewed by Daniel Bates.

Source/WebCore:

Since the deprecatedTargetDPI was an int and the deviceDPI was also an int the result
was truncated.  Changed deprecatedTargetDPI to a float value. Viewport test 91 now passes.

* dom/ViewportArguments.h:
(ViewportArguments):

LayoutTests:

Unskip now passing tests.

* platform/efl/TestExpectations:
* platform/gtk/TestExpectations:
* platform/qt/TestExpectations:

Canonical link: https://commits.webkit.org/108200@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@121635 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/dom/ViewportArguments.h
@@ -110,7 +110,7 @@ struct ViewportArguments {
     }
     // FIXME: We're going to keep this constant around until all embedders
     // refactor their code to no longer need it.
-    static const int deprecatedTargetDPI = 160;
+    static const float deprecatedTargetDPI = 160;
 };
 
 ViewportAttributes computeViewportAttributes(ViewportArguments args, int desktopWidth, int deviceWidth, int deviceHeight, float devicePixelRatio, IntSize visibleViewport);
